backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.11.0
transformers==4.35.0
faiss-cpu==1.13.2
pillow==10.3.0
numpy==1.24.3
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
scipy==1.11.4
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1